PROJECT AIDE/backend/requirements.txt
# === Core Backend Framework ===
fastapi==0.101.1
uvicorn==0.24.0
requests==2.31.0
aiohttp==3.9.5
jinja2==3.1.2
loguru==0.7.0
rich==13.4.2
numpy<2.0  # Pin to NumPy 1.x for compatibility with PyArrow and scikit-learn

# === Code Parsing & Analysis (Repo Agent) ===
tree_sitter==0.20.2  # install via pre-built wheel on Windows if needed
astor==0.8.1
gitpython==3.1.33
pygithub==1.59

# === Vector Database & RAG Framework ===
weaviate-client>=3.20.0,<5.0.0  # Supports both v3 and v4 APIs
langchain>=0.1.172
langchain-core>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.1.0
llama-index>=0.6.18
sentence-transformers>=2.2.2
transformers>=4.39.3
tokenizers>=0.15.2
openai>=0.27.10
faiss-cpu>=1.11.0  # updated to latest version compatible with Python 3.12

# === Document Processing (Docs Agent) ===
markdown2==2.4.2
PyYAML==6.0.3
python-docx>=1.0.0
PyMuPDF==1.25.0
pdfminer.six==20221105
pytesseract==0.3.13
pillow==10.1.0
#pyarrow==10.0.0   # install via pre-built wheel on Windows if needed  # pip install C:\Users\hp\Downloads\pyarrow-11.0.0-cp311-cp311-win_amd64.whl

# === Environment Automation (Setup Agent) ===
docker==6.1.3
#docker-compose>=2.0

# === CI/CD Agent ===
jenkinsapi==0.3.9
celery==5.3.1
redis==5.3.0
psycopg2-binary==2.9.9
pymongo==4.15.3

# === Feedback & Continuous Learning Agent ===
wandb>= 0.16.4
mlflow==2.7.0
dvc==2.42.0

# === Walkthrough / Onboarding Agent ===
sqlite-utils==4.0a0
gTTS==2.4.0
openai-whisper==20250625

# === Visualization / Dashboard ===
streamlit==1.50.0

# === Optional Utilities ===
elastic-apm==6.18.0
python-dotenv==1.1.1